[tools] perf intel-pt: Packet splitting can happen only on 32-bit
Message-id: <20191104222454.10863-129-mpetlan@redhat.com>
Patchwork-id: 282771
O-Subject: [RHEL8.2 BZ1715508 128/327] perf intel-pt: Packet splitting can happen only on 32-bit
Bugzilla: 1715508
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1715508
https://bugzilla.redhat.com/show_bug.cgi?id=1715508

upstream
========
commit 26ee2bcdea33c60aa833cc32a1624ef5d49c9c6f
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Wed Feb 6 12:39:46 2019 +0200

description
===========
Data is copied when the trace is stopped, so packets are never split
between buffers except when processing if the buffer cannot fit in the
address space which can only happen on 32-bit systems. Change the logic
to reflect that.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>