[tools] perf report: Don't shadow inlined symbol with different addr range
Message-id: <20191104222454.10863-164-mpetlan@redhat.com>
Patchwork-id: 282805
O-Subject: [RHEL8.2 BZ1715508 163/327] perf report: Don't shadow inlined symbol with different addr range
Bugzilla: 1715508
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1715508
https://bugzilla.redhat.com/show_bug.cgi?id=1715508

upstream
========
commit 7346195e8643482968f547483e0d823ec1982fab
Author: He Kuang <hekuang@huawei.com>
Date: Tue Feb 19 21:05:31 2019 +0800

description
===========
We can't assume inlined symbols with the same name are equal, because
their address range may be different. This will cause the symbols with
different addresses be shadowed when adding to the hist entry, and lead
to ERANGE error when checking the symbol address during sample parse,
the addr should be within the range of [sym.start, sym.end].

The error message is like: "0x36aea60 [0x8]: failed to process type: 68".

The second parameter of symbol__new() is the length of the fake symbol
for the inline frame, which is the subtraction of the end and start
address of base_sym.

Fixes: aa441895f7b4 ("perf report: Compare symbol name for inlined frames when sorting")
Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>