[tools] perf doc: Fix documentation of the Flags section in perf.data
Message-id: <20191104222454.10863-166-mpetlan@redhat.com>
Patchwork-id: 282809
O-Subject: [RHEL8.2 BZ1715508 165/327] perf doc: Fix documentation of the Flags section in perf.data
Bugzilla: 1715508
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1715508
https://bugzilla.redhat.com/show_bug.cgi?id=1715508

upstream
========
commit 8c23a522388b34cd7bc8473987eda0c75eb37c0e
Author: Jonas Rabenstein <jonas.rabenstein@studium.uni-erlangen.de>
Date: Tue Feb 19 16:45:15 2019 +0100

description
===========
According to the current documentation the flags section is placed after
the file header itself but the code assumes to find the flags section
after the data section. This change updates the documentation to that
assumption.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>