[tools] perf scripts python: exported-sql-viewer.py: Remove no selection error
Message-id: <20191104222454.10863-186-mpetlan@redhat.com>
Patchwork-id: 282829
O-Subject: [RHEL8.2 BZ1715508 185/327] perf scripts python: exported-sql-viewer.py: Remove no selection error
Bugzilla: 1715508
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1715508
https://bugzilla.redhat.com/show_bug.cgi?id=1715508

upstream
========
commit fc2c77aa8437855d2992d3f3c6a1dff681789a07
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Fri Feb 22 09:27:27 2019 +0200

description
===========
If no selection is made on the 'Selected branches' dialog, then the
output is the same as the 'All branches' report. That is not really an
error, and is not desirable for future reports, so remove it.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>