[tools] perf db-export: Add calls parent_id to enable creation of call trees
Message-id: <20191104222454.10863-206-mpetlan@redhat.com>
Patchwork-id: 282846
O-Subject: [RHEL8.2 BZ1715508 205/327] perf db-export: Add calls parent_id to enable creation of call trees
Bugzilla: 1715508
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1715508
https://bugzilla.redhat.com/show_bug.cgi?id=1715508

upstream
========
commit f435887ec0c941b97301bd6ed1f3e4b5200df409
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Thu Feb 28 15:00:24 2019 +0200

description
===========
The call_path can be used to find the parent symbol for a call but not
the exact parent call. To do that add parent_id to the call_return
export. This enables the creation of a call tree from the exported data.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>