[tools] perf data: Don't store auxtrace index for directory data file
Message-id: <20191104222454.10863-235-mpetlan@redhat.com>
Patchwork-id: 282878
O-Subject: [RHEL8.2 BZ1715508 234/327] perf data: Don't store auxtrace index for directory data file
Bugzilla: 1715508
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1715508
https://bugzilla.redhat.com/show_bug.cgi?id=1715508

upstream
========
commit cd3dd8dd8ff62374d90cb3f2e54b8c94106c7810
Author: Jiri Olsa <jolsa@kernel.org>
Date: Fri Mar 8 14:47:36 2019 +0100

description
===========
We can't store the auxtrace index when we store into multiple files,
because we keep only offset for it, not the file.

The auxtrace data will be processed correctly in the 'pipe' mode.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>