[tools] perf feature detection: Add -lopcodes to feature-libbfd
Message-id: <20191104222454.10863-284-mpetlan@redhat.com>
Patchwork-id: 282927
O-Subject: [RHEL8.2 BZ1715508 283/327] perf feature detection: Add -lopcodes to feature-libbfd
Bugzilla: 1715508
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1715508
https://bugzilla.redhat.com/show_bug.cgi?id=1715508

upstream
========
commit 31be9478ed7f43d6351e0d5a2257ca76609c83d3
Author: Song Liu <songliubraving@fb.com>
Date: Mon Mar 11 22:30:47 2019 -0700

description
===========
Both libbfd and libopcodes are distributed with binutil-dev/devel. When
libbfd is present, it is OK to assume that libopcodes also present. This
has been a safe assumption for bpftool.

This patch adds -lopcodes to perf/Makefile.config. libopcodes will be
used in the next commit for BPF annotation.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>