[tools] tools lib traceevent: Fix missing equality check for strcmp
Message-id: <20191104222454.10863-306-mpetlan@redhat.com>
Patchwork-id: 282955
O-Subject: [RHEL8.2 BZ1715508 305/327] tools lib traceevent: Fix missing equality check for strcmp
Bugzilla: 1715508
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1715508
https://bugzilla.redhat.com/show_bug.cgi?id=1715508

upstream
========
commit f32c2877bcb068a718bb70094cd59ccc29d4d082
Author: Rikard Falkeborn <rikard.falkeborn@gmail.com>
Date: Tue Apr 9 11:15:29 2019 +0200

description
===========
There was a missing comparison with 0 when checking if type is "s64" or
"u64". Therefore, the body of the if-statement was entered if "type" was
"u64" or not "s64", which made the first strcmp() redundant since if
type is "u64", it's not "s64".

If type is "s64", the body of the if-statement is not entered but since
the remainder of the function consists of if-statements which will not
be entered if type is "s64", we will just return "val", which is
correct, albeit at the cost of a few more calls to strcmp(), i.e., it
will behave just as if the if-statement was entered.

If type is neither "s64" or "u64", the body of the if-statement will be
entered incorrectly and "val" returned. This means that any type that is
checked after "s64" and "u64" is handled the same way as "s64" and
"u64", i.e., the limiting of "val" to fit in for example "s8" is never
reached.

This was introduced in the kernel tree when the sources were copied from
trace-cmd in commit f7d82350e597 ("tools/events: Add files to create
libtraceevent.a"), and in the trace-cmd repo in 1cdbae6035cei
("Implement typecasting in parser") when the function was introduced,
i.e., it has always behaved the wrong way.

Detected by cppcheck.

Fixes: f7d82350e597 ("tools/events: Add files to create libtraceevent.a")
Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>