[tools] tools lib traceevent: Change tag string for error
Message-id: <20191104222454.10863-317-mpetlan@redhat.com>
Patchwork-id: 282958
O-Subject: [RHEL8.2 BZ1715508 316/327] tools lib traceevent: Change tag string for error
Bugzilla: 1715508
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1715508
https://bugzilla.redhat.com/show_bug.cgi?id=1715508

upstream
========
commit 5f05182fab9a29fea6c4ab8113be45adf0c11bf0
Author: Leo Yan <leo.yan@linaro.org>
Date: Wed Apr 24 09:38:02 2019 +0800

description
===========
The traceevent lib is used by the perf tool, and when executing

  perf test -v 6

it outputs error log on the ARM64 platform:

  running test 33 '*:*'trace-cmd: No such file or directory

  [...]

  trace-cmd: Invalid argument

The trace event parsing code originally came from trace-cmd so it keeps
the tag string "trace-cmd" for errors, this easily introduces the
impression that the perf tool launches trace-cmd command for trace event
parsing, but in fact the related parsing is accomplished by the
traceevent lib.

This patch changes the tag string to "libtraceevent" so that we can
avoid confusion and let users to more easily connect the error with
traceevent lib.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>