[tools] perf bench numa: Add define for RUSAGE_THREAD if not present
Message-id: <20191104222454.10863-318-mpetlan@redhat.com>
Patchwork-id: 282965
O-Subject: [RHEL8.2 BZ1715508 317/327] perf bench numa: Add define for RUSAGE_THREAD if not present
Bugzilla: 1715508
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1715508
https://bugzilla.redhat.com/show_bug.cgi?id=1715508

upstream
========
commit bf561d3c13423fc54daa19b5d49dc15fafdb7acc
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Thu Apr 25 18:36:51 2019 -0300

description
===========
While cross building perf to the ARC architecture on a fedora 30 host,
we were failing with:

      CC       /tmp/build/perf/bench/numa.o
  bench/numa.c: In function worker_thread:
  bench/numa.c:1261:12: error: RUSAGE_THREAD undeclared (first use in this function); did you mean SIGEV_THREAD?
    getrusage(RUSAGE_THREAD, &rusage);
              ^~~~~~~~~~~~~
              SIGEV_THREAD
  bench/numa.c:1261:12: note: each undeclared identifier is reported only once for each function it appears in

[perfbuilder@60d5802468f6 perf]$ /arc_gnu_2019.03-rc1_prebuilt_uclibc_le_archs_linux_install/bin/arc-linux-gcc --version | head -1
arc-linux-gcc (ARCv2 ISA Linux uClibc toolchain 2019.03-rc1) 8.3.1 20190225
[perfbuilder@60d5802468f6 perf]$

Trying to reproduce a report by Vineet, I noticed that, with just
cross-built zlib and numactl libraries, I ended up with the above
failure.

So, since RUSAGE_THREAD is available as a define, check for that and
numactl libraries, I ended up with the above failure.

So, since RUSAGE_THREAD is available as a define in the system headers,
check if it is defined in the 'perf bench numa' sources and define it if
not.

Now it builds and I have to figure out if the problem reported by Vineet
only takes place if we have libelf or some other library available.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>