[tools] perf evsel: Do not rely on errno values for precise_ip fallback
Message-id: <20191104222454.10863-328-mpetlan@redhat.com>
Patchwork-id: 282971
O-Subject: [RHEL8.2 BZ1715508 327/327] perf evsel: Do not rely on errno values for precise_ip fallback
Bugzilla: 1715508
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1715508
https://bugzilla.redhat.com/show_bug.cgi?id=1715508

upstream
========
commit cd136189370cc8a5aec0ea4b4ec517e5ee38d8a0
Author: Jiri Olsa <jolsa@kernel.org>
Date:   Wed Jul 3 10:09:49 2019 +0200

description
===========
Konstantin reported problem with default perf record command, which
fails on some AMD servers, because of the default maximum precise
config.

The current fallback mechanism counts on getting ENOTSUP errno for
precise_ip fails, but that's not the case on some AMD servers.

We can fix this by removing the errno check completely, because the
precise_ip fallback is separated. We can just try  (if requested by
evsel->precise_max) all possible precise_ip, and if one succeeds we win,
if not, we continue with standard fallback.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>