[arm] arm_pmu: Tidy up clear_event_idx call backs
Message-id: <20191107124025.28024-3-eric.auger@redhat.com>
Patchwork-id: 283791
O-Subject: [RHEL8.2 virt PATCH v4 02/75] arm_pmu: Tidy up clear_event_idx call backs
Bugzilla: 1749501
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Mark Salter <msalter@redhat.com>
RH-Acked-by: Andrew Jones <drjones@redhat.com>

From: Suzuki K Poulose <suzuki.poulose@arm.com>

The armpmu uses get_event_idx callback to allocate an event
counter for a given event, which marks the selected counter
as "used". Now, when we delete the counter, the arm_pmu goes
ahead and clears the "used" bit and then invokes the "clear_event_idx"
call back, which kind of splits the job between the core code
and the backend. To keep things tidy, mandate the implementation
of clear_event_idx() and add it for exisiting backends.
This will be useful for adding the chained event support, where
we leave the event idx maintenance to the backend.

Also, when an event is removed from the PMU, reset the hw.idx
to indicate that a counter is not allocated for this event,
to help the backends do better checks. This will be also used
for the chain counter support.

Cc: Will Deacon <will.deacon@arm.com>
Cc: Mark Rutland <mark.rutland@arm.com>
Reviewed-by: Julien Thierry <julien.thierry@arm.com>
Signed-off-by: Suzuki K Poulose <suzuki.poulose@arm.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>
(cherry picked from commit 7dfc8db1d117ae08c649266d5459ac8b10800d6e)
Signed-off-by: Eric Auger <eric.auger@redhat.com>
Signed-off-by: Phillip Lougher <plougher@redhat.com>