[arm64] kvm: arm64: Migrate _elx sysreg accessors to msr_s/mrs_s
Message-id: <20191107124025.28024-41-eric.auger@redhat.com>
Patchwork-id: 283826
O-Subject: [RHEL8.2 virt PATCH v4 40/75] KVM: arm64: Migrate _elx sysreg accessors to msr_s/mrs_s
Bugzilla: 1749501
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Mark Salter <msalter@redhat.com>
RH-Acked-by: Andrew Jones <drjones@redhat.com>

From: Dave Martin <Dave.Martin@arm.com>

Currently, the {read,write}_sysreg_el*() accessors for accessing
particular ELs' sysregs in the presence of VHE rely on some local
hacks and define their system register encodings in a way that is
inconsistent with the core definitions in <asm/sysreg.h>.

As a result, it is necessary to add duplicate definitions for any
system register that already needs a definition in sysreg.h for
other reasons.

This is a bit of a maintenance headache, and the reasons for the
_el*() accessors working the way they do is a bit historical.

This patch gets rid of the shadow sysreg definitions in
<asm/kvm_hyp.h>, converts the _el*() accessors to use the core
__msr_s/__mrs_s interface, and converts all call sites to use the
standard sysreg #define names (i.e., upper case, with SYS_ prefix).

This patch will conflict heavily anyway, so the opportunity
to clean up some bad whitespace in the context of the changes is
taken.

The change exposes a few system registers that have no sysreg.h
definition, due to msr_s/mrs_s being used in place of msr/mrs:
additions are made in order to fill in the gaps.

Signed-off-by: Dave Martin <Dave.Martin@arm.com>
Cc: Catalin Marinas <catalin.marinas@arm.com>
Cc: Christoffer Dall <christoffer.dall@arm.com>
Cc: Mark Rutland <mark.rutland@arm.com>
Cc: Will Deacon <will.deacon@arm.com>
Link: https://www.spinics.net/lists/kvm-arm/msg31717.html
[Rebased to v4.21-rc1]
Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>
[Rebased to v5.2-rc5, changelog updates]
Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
(cherry picked from commit fdec2a9ef853172529baaa192673b4cdb9a44fac)
Signed-off-by: Eric Auger <eric.auger@redhat.com>
Signed-off-by: Phillip Lougher <plougher@redhat.com>