[virt] kvm: arm/arm64: vgic: Reevaluate level sensitive interrupts on enable
Message-id: <20191107124025.28024-51-eric.auger@redhat.com>
Patchwork-id: 283837
O-Subject: [RHEL8.2 virt PATCH v4 50/75] KVM: arm/arm64: vgic: Reevaluate level sensitive interrupts on enable
Bugzilla: 1749501
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Mark Salter <msalter@redhat.com>
RH-Acked-by: Andrew Jones <drjones@redhat.com>

From: Alexandru Elisei <alexandru.elisei@arm.com>

A HW mapped level sensitive interrupt asserted by a device will not be put
into the ap_list if it is disabled at the VGIC level. When it is enabled
again, it will be inserted into the ap_list and written to a list register
on guest entry regardless of the state of the device.

We could argue that this can also happen on real hardware, when the command
to enable the interrupt reached the GIC before the device had the chance to
de-assert the interrupt signal; however, we emulate the distributor and
redistributors in software and we can do better than that.

Signed-off-by: Alexandru Elisei <alexandru.elisei@arm.com>
Signed-off-by: Marc Zyngier <maz@kernel.org>
(cherry picked from commit 16e604a437c89751dc626c9e90cf88ba93c5be64)
Signed-off-by: Eric Auger <eric.auger@redhat.com>
Signed-off-by: Phillip Lougher <plougher@redhat.com>