[virt] kvm: arm/arm64: vgic-its: Invalidate MSI-LPI translation cache on specific commands
Message-id: <20191107124025.28024-60-eric.auger@redhat.com>
Patchwork-id: 283844
O-Subject: [RHEL8.2 virt PATCH v4 59/75] KVM: arm/arm64: vgic-its: Invalidate MSI-LPI translation cache on specific commands
Bugzilla: 1749501
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Mark Salter <msalter@redhat.com>
RH-Acked-by: Andrew Jones <drjones@redhat.com>

From: Marc Zyngier <maz@kernel.org>

The LPI translation cache needs to be discarded when an ITS command
may affect the translation of an LPI (DISCARD, MAPC and MAPD with V=0)
or the routing of an LPI to a redistributor with disabled LPIs (MOVI,
MOVALL).

We decide to perform a full invalidation of the cache, irrespective
of the LPI that is affected. Commands are supposed to be rare enough
that it doesn't matter.

Tested-by: Andre Przywara <andre.przywara@arm.com>
Reviewed-by: Eric Auger <eric.auger@redhat.com>
Signed-off-by: Marc Zyngier <maz@kernel.org>
(cherry picked from commit 0c14484866194382c0631d6e9bb14cc2fe83e814)
Signed-off-by: Eric Auger <eric.auger@redhat.com>
Signed-off-by: Phillip Lougher <plougher@redhat.com>