[cpuidle] cpuidle: menu: Fix wakeup statistics updates for polling state
Message-id: <20191107134746.065311262@amt.cnet>
Patchwork-id: 283888
O-Subject: [RHEL 8.2 PATCH 03/18] cpuidle: menu: Fix wakeup statistics updates for polling state
Bugzilla: 1759282
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Vitaly Kuznetsov <vkuznets@redhat.com>
RH-Acked-by: Andrea Arcangeli <aarcange@redhat.com>

[-RHEL backport note: use RH_KABI_FILL_HOLE to avoid kABI breakage]

# pahole -C cpuidle_device drivers/cpuidle/cpuidle.o
struct cpuidle_device {
	unsigned int               registered:1;         /*     0: 0  4 */
	unsigned int               enabled:1;            /*     0: 1  4 */
	unsigned int               use_deepest_state:1;  /*     0: 2  4 */

	/* XXX 29 bits hole, try to pack */

	unsigned int               cpu;                  /*     4     4 */
	int                        last_residency;       /*     8     4 */

	/* XXX 4 bytes hole, try to pack */

	struct cpuidle_state_usage states_usage[10];     /*    16   400 */
	/* --- cacheline 6 boundary (384 bytes) was 32 bytes ago --- */
	struct cpuidle_state_kobj * kobjs[10];           /*   416    80 */
	/* --- cacheline 7 boundary (448 bytes) was 48 bytes ago --- */
	struct cpuidle_driver_kobj * kobj_driver;        /*   496     8 */
	struct cpuidle_device_kobj * kobj_dev;           /*   504     8 */
	/* --- cacheline 8 boundary (512 bytes) --- */
	struct list_head   device_list;                  /*   512    16 */

	/* size: 528, cachelines: 9, members: 10 */
	/* sum members: 520, holes: 1, sum holes: 4 */
	/* sum bitfield members: 3 bits, bit holes: 1, sum bit holes: 29 bits */
	/* last cacheline: 16 bytes */
};

The cpuidle structures and API is not kABI white listed, therefore external
users are unsupported.

The chain which breaks "struct task_struct" is:

task_struct -> sched_class -> rq -> idle_state -> cpuidle_device

BZ: 1759282

commit 5f26bdceb9c0a5e6c696aa2899d077cd3ae93413

If the CPU exits the "polling" state due to the time limit in the
loop in poll_idle(), this is not a real wakeup and it just means
that the "polling" state selection was not adequate.  The governor
mispredicted short idle duration, but had a more suitable state been
selected, the CPU might have spent more time in it.  In fact, there
is no reason to expect that there would have been a wakeup event
earlier than the next timer in that case.

Handling such cases as regular wakeups in menu_update() may cause the
menu governor to make suboptimal decisions going forward, but ignoring
them altogether would not be correct either, because every time
menu_select() is invoked, it makes a separate new attempt to predict
the idle duration taking distinct time to the closest timer event as
input and the outcomes of all those attempts should be recorded.

For this reason, make menu_update() always assume that if the
"polling" state was exited due to the time limit, the next proper
wakeup event for the CPU would be the next timer event (not
including the tick).

Fixes: a37b969a61c1 "cpuidle: poll_state: Add time limit to poll_idle()"
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Acked-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Reviewed-by: Daniel Lezcano <daniel.lezcano@linaro.org>
Signed-off-by: Marcelo Tosatti <mtosatti@redhat.com>
Signed-off-by: Phillip Lougher <plougher@redhat.com>