[cpuidle] cpuidle: poll_state: Disregard disable idle states
Message-id: <20191107134746.243422680@amt.cnet>
Patchwork-id: 283877
O-Subject: [RHEL 8.2 PATCH 05/18] cpuidle: poll_state: Disregard disable idle states
Bugzilla: 1759282
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Vitaly Kuznetsov <vkuznets@redhat.com>
RH-Acked-by: Andrea Arcangeli <aarcange@redhat.com>

BZ: 1759282

commit 800fb34a99ce7d22dca839c90f869c7a12b50f70

When computing the limit of time to spend in the loop in poll_idle(),
use the target residency of the first enabled idle state deeper than
state 0 instead of always using the target residency of state 1.

This helps when state 1 is disabled for diagnostics, for instance.

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Marcelo Tosatti <mtosatti@redhat.com>
Signed-off-by: Phillip Lougher <plougher@redhat.com>