[cpuidle] cpuidle: allow governor switch on cpuidle_register_driver()
Message-id: <20191107134747.025296800@amt.cnet>
Patchwork-id: 283872
O-Subject: [RHEL 8.2 PATCH 14/18] cpuidle: allow governor switch on cpuidle_register_driver()
Bugzilla: 1759282
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Vitaly Kuznetsov <vkuznets@redhat.com>
RH-Acked-by: Andrea Arcangeli <aarcange@redhat.com>

-RHEL backport note: this upstream patch adds a pointer
to "struct cpuidle_driver". The interface is not kABI whitelisted,
but unfortunately attempting to add a pointer with RH_KABI_EXTEND
changes the checksum of "struct task_struct *" via:

task_struct -> sched_class -> rq -> cpuidle_state -> cpuidle_driver.

There is no hole large enough for "struct cpuidle_driver", so we add
a new function rhel_cpuidle_register_driver_hpoll, passing
the governor directly. That function is limited to work with
the halt poll driver only.
There should be no more drivers that use the "governor" field
in the near future anyway (and there is no interest in an
external cpuidle driver).

BZ: 1759282

commit cb5d8c45ab6c3daf8269e550cfb2d5018a876fe3

The recently introduced haltpoll driver is largely only useful with
haltpoll governor. To allow drivers to associate with a particular idle
behaviour, add a @governor property to 'struct cpuidle_driver' and thus
allow a cpuidle driver to switch to a *preferred* governor on idle driver
registration. We save the previous governor, and when an idle driver is
unregistered we switch back to that.

The @governor can be overridden by cpuidle.governor= boot param or
alternatively be ignored if the governor doesn't exist.

Signed-off-by: Joao Martins <joao.m.martins@oracle.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Phillip Lougher <plougher@redhat.com>