[cpuidle] cpuidle-haltpoll: do not set an owner to allow modunload
Message-id: <20191107134747.174028570@amt.cnet>
Patchwork-id: 283876
O-Subject: [RHEL 8.2 PATCH 16/18] cpuidle-haltpoll: do not set an owner to allow modunload
Bugzilla: 1759282
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Vitaly Kuznetsov <vkuznets@redhat.com>
RH-Acked-by: Andrea Arcangeli <aarcange@redhat.com>

BZ: 1759282

commit 472f263660832b90e53bede2020f68cd14f8b76c

cpuidle-haltpoll can be built as a module to allow optional late load.
Given we are setting @owner to THIS_MODULE, cpuidle will attempt to grab a
module reference every time a cpuidle_device is registered -- so
essentially all online cpus get a reference.

This prevents for the module to be unloaded later, which makes the
module_exit callback entirely unused. Thus remove the @owner and allow
module to be unloaded.

Fixes: fa86ee90eb11 ("add cpuidle-haltpoll driver")
Signed-off-by: Joao Martins <joao.m.martins@oracle.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Marcelo Tosatti <mtosatti@redhat.com>
Signed-off-by: Phillip Lougher <plougher@redhat.com>