[net] vsock: correct removal of socket from the list
Message-id: <20191107140611.42235-6-sgarzare@redhat.com>
Patchwork-id: 283897
O-Subject: [RHEL8.2 PATCH v2 5/6] vsock: correct removal of socket from the list
Bugzilla: 1769714
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>
RH-Acked-by: Cathy Avery <cavery@redhat.com>

From: Sunil Muthuswamy <sunilmut@microsoft.com>

The current vsock code for removal of socket from the list is both
subject to race and inefficient. It takes the lock, checks whether
the socket is in the list, drops the lock and if the socket was on the
list, deletes it from the list. This is subject to race because as soon
as the lock is dropped once it is checked for presence, that condition
cannot be relied upon for any decision. It is also inefficient because
if the socket is present in the list, it takes the lock twice.

Signed-off-by: Sunil Muthuswamy <sunilmut@microsoft.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit d5afa82c977ea06f7119058fa0eb8519ea501031)
Signed-off-by: Stefano Garzarella <sgarzare@redhat.com>
Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>