[fs] ceph: handle the case where a dentry has been renamed on outstanding req
Message-id: <20191108140453.73236-106-jlayton@redhat.com>
Patchwork-id: 284279
O-Subject: [BZ#1767959 RHEL8.2 PATCH 105/264] ceph: handle the case where a dentry has been renamed on outstanding req
Bugzilla: 1767959
RH-Acked-by: Zheng Yan <zyan@redhat.com>
RH-Acked-by: Ilya Dryomov <idryomov@redhat.com>

From: Jeff Layton <jlayton@kernel.org>

It's possible for us to issue a lookup to revalidate a dentry
concurrently with a rename. If done in the right order, then we could
end up processing dentry info in the reply that no longer reflects the
state of the dentry.

If req->r_dentry->d_name differs from the one in the trace, then just
ignore the trace in the reply. We only need to do this however if the
parent's i_rwsem is not held.

Signed-off-by: Jeff Layton <jlayton@kernel.org>
Reviewed-by: "Yan, Zheng" <zyan@redhat.com>
Signed-off-by: Ilya Dryomov <idryomov@gmail.com>
(cherry picked from commit 4b8222870032715f9d995f3eb7c7acd8379a275d)
Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>