[x86] perf/x86/lbr: Avoid reading the LBRs when adaptive PEBS handles them
Message-id: <20191110211921.28070-17-mpetlan@redhat.com>
Patchwork-id: 285022
O-Subject: [RHEL8.2 BZ1743208 016/158] perf/x86/lbr: Avoid reading the LBRs when adaptive PEBS handles them
Bugzilla: 1743208
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1743208
https://bugzilla.redhat.com/show_bug.cgi?id=1743208

upstream
========
commit d3617b98b04583df222f34992e65712862a77bf1
Author: Andi Kleen <ak@linux.intel.com>
Date: Tue Apr 2 12:45:03 2019 -0700

description
===========
With adaptive PEBS the CPU can directly supply the LBR information,
so we don't need to read it again. But the LBRs still need to be
enabled. Add a special count to the cpuc that distinguishes these
two cases, and avoid reading the LBRs unnecessarily when PEBS is
active.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>