[x86] perf/x86: Support constraint ranges
Message-id: <20191110211921.28070-18-mpetlan@redhat.com>
Patchwork-id: 285023
O-Subject: [RHEL8.2 BZ1743208 017/158] perf/x86: Support constraint ranges
Bugzilla: 1743208
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1743208
https://bugzilla.redhat.com/show_bug.cgi?id=1743208

upstream
========
commit 63b79f6ebc464afb730bc45762c820795e276da1
Author: Peter Zijlstra <peterz@infradead.org>
Date: Tue Apr 2 12:45:04 2019 -0700

description
===========
Icelake extended the general counters to 8, even when SMT is enabled.
However only a (large) subset of the events can be used on all 8
counters.

The events that can or cannot be used on all counters are organized
in ranges.

A lot of scheduler constraints are required to handle all this.

To avoid blowing up the tables add event code ranges to the constraint
tables, and a new inline function to match them.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>