[x86] perf/x86/intel: Fix INTEL_FLAGS_EVENT_CONSTRAINT* masking
Message-id: <20191110211921.28070-26-mpetlan@redhat.com>
Patchwork-id: 285028
O-Subject: [RHEL8.2 BZ1743208 025/158] perf/x86/intel: Fix INTEL_FLAGS_EVENT_CONSTRAINT* masking
Bugzilla: 1743208
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1743208
https://bugzilla.redhat.com/show_bug.cgi?id=1743208

upstream
========
commit 6b89d4c1ae8596a8c9240f169ef108704de373f2
Author: Stephane Eranian <eranian@google.com>
Date: Thu May 9 14:45:56 2019 -0700

description
===========
On Intel Westmere, a cmdline as follows:

  $ perf record -e cpu/event=0xc4,umask=0x2,name=br_inst_retired.near_call/p ....

was failing. Yet the event+ umask support PEBS.

It turns out this is due to a bug in the the PEBS event constraint table for
westmere. All forms of BR_INST_RETIRED.* support PEBS. Therefore the constraint
mask should ignore the umask. The name of the macro INTEL_FLAGS_EVENT_CONSTRAINT()
hint that this is the case but it was not. That macros was checking both the
event code and event umask. Therefore, it was only matching on 0x00c4.
There are code+umask macros, they all have *UEVENT*.

This bug fixes the issue by checking only the event code in the mask.
Both single and range version are modified.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>