[kernel] perf/ring_buffer: Fix exposing a temporarily decreased data_head
Message-id: <20191110211921.28070-29-mpetlan@redhat.com>
Patchwork-id: 285032
O-Subject: [RHEL8.2 BZ1743208 028/158] perf/ring_buffer: Fix exposing a temporarily decreased data_head
Bugzilla: 1743208
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1743208
https://bugzilla.redhat.com/show_bug.cgi?id=1743208

upstream
========
commit 1b038c6e05ff70a1e66e3e571c2e6106bdb75f53
Author: Yabin Cui <yabinc@google.com>
Date: Fri May 17 13:52:31 2019 +0200

description
===========
In perf_output_put_handle(), an IRQ/NMI can happen in below location and
write records to the same ring buffer:

	...
	local_dec_and_test(&rb->nest)
	...                          <-- an IRQ/NMI can happen here
	rb->user_page->data_head = head;
	...

In this case, a value A is written to data_head in the IRQ, then a value
B is written to data_head after the IRQ. And A > B. As a result,
data_head is temporarily decreased from A to B. And a reader may see
data_head < data_tail if it read the buffer frequently enough, which
creates unexpected behaviors.

This can be fixed by moving dec(&rb->nest) to after updating data_head,
which prevents the IRQ/NMI above from updating data_head.

[ Split up by peterz. ]

Fixes: ef60777c9abd ("perf: Optimize the perf_output() path by removing IRQ-disables")
Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>