[tools] tools lib traceevent: Handle trace_printk() "%px"
Message-id: <20191110211921.28070-49-mpetlan@redhat.com>
Patchwork-id: 285053
O-Subject: [RHEL8.2 BZ1743208 048/158] tools lib traceevent: Handle trace_printk() "%px"
Bugzilla: 1743208
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1743208
https://bugzilla.redhat.com/show_bug.cgi?id=1743208

upstream
========
commit 328b82b7497787cf4e4856a216dbf4e8a54c06a6
Author: Steven Rostedt (VMware) <rostedt@goodmis.org>
Date: Mon Apr 1 12:43:06 2019 -0400

description
===========
With security updates, %p in the kernel is hashed to protect true kernel
locations. But trace_printk() is not allowed in production systems, and
when a pointer is used, there are many times that the actual address is
needed. "%px" produces the real address. But libtraceevent does not know how
to handle that extension. Add it.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>