[tools] tools lib traceevent: Implement a new API, tep_list_events_copy()
Message-id: <20191110211921.28070-51-mpetlan@redhat.com>
Patchwork-id: 285054
O-Subject: [RHEL8.2 BZ1743208 050/158] tools lib traceevent: Implement a new API, tep_list_events_copy()
Bugzilla: 1743208
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1743208
https://bugzilla.redhat.com/show_bug.cgi?id=1743208

upstream
========
commit 6699ed712a97f70267e017f509126b890f6f6b28
Author: Tzvetomir Stoyanov <tstoyanov@vmware.com>
Date: Mon Apr 1 12:43:08 2019 -0400

description
===========
Existing API tep_list_events() is not thread safe, it uses the internal
array sort_events to keep cache of the sorted events and reuses it. This
patch implements a new API, tep_list_events_copy(), which allocates new
sorted array each time it is called. It could be used when a sorted
events functionality is needed in thread safe use cases. It is up to the
caller to free the array.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>