[tools] tools lib traceevent: Remove call to exit() from tep_filter_add_filter_str()
Message-id: <20191110211921.28070-58-mpetlan@redhat.com>
Patchwork-id: 285061
O-Subject: [RHEL8.2 BZ1743208 057/158] tools lib traceevent: Remove call to exit() from tep_filter_add_filter_str()
Bugzilla: 1743208
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1743208
https://bugzilla.redhat.com/show_bug.cgi?id=1743208

upstream
========
commit fea6b632235b9bedc58c72cd24f1865bb0c365db
Author: Tzvetomir Stoyanov <tstoyanov@vmware.com>
Date: Mon Apr 1 12:43:15 2019 -0400

description
===========
This patch removes call to exit() from tep_filter_add_filter_str(). A
library function should not force the application to exit. In the
current implementation tep_filter_add_filter_str() calls exit() when a
special "test_filters" mode is set, used only for debugging purposes.
When this mode is set and a filter is added - its string is printed to
the console and exit() is called. This patch changes the logic - when in
"test_filters" mode, the filter string is still printed, but the exit()
is not called. It is up to the application to track when "test_filters"
mode is set and to call exit, if needed.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>