[tools] perf report: Add stub processing of compressed events for -D
Message-id: <20191110211921.28070-106-mpetlan@redhat.com>
Patchwork-id: 285109
O-Subject: [RHEL8.2 BZ1743208 105/158] perf report: Add stub processing of compressed events for -D
Bugzilla: 1743208
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1743208
https://bugzilla.redhat.com/show_bug.cgi?id=1743208

upstream
========
commit 61a7773ca88f32ef7e185fdf9fc0d44e8ec18a66
Author: Alexey Budankov <alexey.budankov@linux.intel.com>
Date: Mon Mar 18 20:45:11 2019 +0300

description
===========
Committer note:

Split from a larger patch, this only dumps PERF_RECORD_COMPRESSED as
unhandled, so that when we introduce the record part in the next patch,
we don't see unhandled events when using 'perf record -D'.

Changed it so that we dump the event if the handler is just a stub, i.e.
for the case where we don't have ZSTD linked but we're processing a
perf.data file generated by a tool with that linked.

Also when failing to decompress we can't just dump the uncompressed
event and return 0, we have to propagate the error.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>