[tools] perf intel-pt: Fix improved sample timestamp
Message-id: <20191110211921.28070-120-mpetlan@redhat.com>
Patchwork-id: 285122
O-Subject: [RHEL8.2 BZ1743208 119/158] perf intel-pt: Fix improved sample timestamp
Bugzilla: 1743208
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1743208
https://bugzilla.redhat.com/show_bug.cgi?id=1743208

upstream
========
commit 61b6e08dc8e3ea80b7485c9b3f875ddd45c8466b
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Fri May 10 15:41:42 2019 +0300

description
===========
The decoder uses its current timestamp in samples. Usually that is a
timestamp that has already passed, but in some cases it is a timestamp
for a branch that the decoder is walking towards, and consequently
hasn't reached.

The intel_pt_sample_time() function decides which is which, but was not
handling TNT packets exactly correctly.

In the case of TNT, the timestamp applies to the first branch, so the
decoder must first walk to that branch.

That means intel_pt_sample_time() should return true for TNT, and this
patch makes that change. However, if the first branch is a non-taken
branch (i.e. a 'N'), then intel_pt_sample_time() needs to return false
for subsequent taken branches in the same TNT packet.

To handle that, introduce a new state INTEL_PT_STATE_TNT_CONT to
distinguish the cases.

Note that commit 3f04d98e972b5 ("perf intel-pt: Improve sample
timestamp") was also a stable fix and appears, for example, in v4.4
stable tree as commit a4ebb58fd124 ("perf intel-pt: Improve sample
timestamp").

Fixes: 3f04d98e972b5 ("perf intel-pt: Improve sample timestamp")
Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>