[tools] perf docs: Add description for stderr
Message-id: <20191110211921.28070-122-mpetlan@redhat.com>
Patchwork-id: 285124
O-Subject: [RHEL8.2 BZ1743208 121/158] perf docs: Add description for stderr
Bugzilla: 1743208
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1743208
https://bugzilla.redhat.com/show_bug.cgi?id=1743208

upstream
========
commit 6cf626563998d3f8770cc25146986810ee4a5969
Author: Thomas Richter <tmricht@linux.ibm.com>
Date: Mon May 13 10:02:20 2019 +0200

description
===========
'perf report' displays recorded data on the screen and emits warnings
and debug messages in the status line (last one on screen).

perf also supports the possibility to write all debug messages to stderr
(instead of writing them to the status line).

This is achieved with the following command:

  # ./perf --debug stderr=1 report -vvvvv -i ~/fast.data 2>/tmp/2
  # ll /tmp/2
  -rw-rw-r-- 1 tmricht tmricht 5420835 May  7 13:46 /tmp/2
  #

The usage of variable stderr=1 is not documented, so add it to the perf
man page.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>