[tools] perf test vmlinux-kallsyms: Ignore aliases to _etext when searching on kallsyms
Message-id: <20191110211921.28070-129-mpetlan@redhat.com>
Patchwork-id: 285132
O-Subject: [RHEL8.2 BZ1743208 128/158] perf test vmlinux-kallsyms: Ignore aliases to _etext when searching on kallsyms
Bugzilla: 1743208
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1743208
https://bugzilla.redhat.com/show_bug.cgi?id=1743208

upstream
========
commit 93f678b9ae8feb7f7cec29c2dcbb5756ad22d9a1
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Fri May 24 15:39:00 2019 -0300

description
===========
No need to search for aliases for the symbol that marks the end of the
kernel text segment, the following patch will make such symbols not to
be found when searching in the kallsyms maps causing this test to fail.

So as a prep patch to avoid breaking bisection, ignore such symbols.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>