[net] vsock: Send reset control packet when socket is partially bound
Message-id: <20191111173136.52465-2-sgarzare@redhat.com>
Patchwork-id: 285274
O-Subject: [RHEL8.2 PATCH 1/2] VSOCK: Send reset control packet when socket is partially bound
Bugzilla: 1769904
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>
RH-Acked-by: Cathy Avery <cavery@redhat.com>

From: Jorgen Hansen <jhansen@vmware.com>

If a server side socket is bound to an address, but not in the listening
state yet, incoming connection requests should receive a reset control
packet in response. However, the function used to send the reset
silently drops the reset packet if the sending socket isn't bound
to a remote address (as is the case for a bound socket not yet in
the listening state). This change fixes this by using the src
of the incoming packet as destination for the reset packet in
this case.

Fixes: d021c344051a ("VSOCK: Introduce VM Sockets")
Reviewed-by: Adit Ranadive <aditr@vmware.com>
Reviewed-by: Vishnu Dasa <vdasa@vmware.com>
Signed-off-by: Jorgen Hansen <jhansen@vmware.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit a915b982d8f5e4295f64b8dd37ce753874867e88)
Signed-off-by: Stefano Garzarella <sgarzare@redhat.com>
Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>