[infiniband] IB/{hfi1, qib, rdmavt}: Put qp in error state when cq is full
Message-id: <20191112031053.14620-42-honli@redhat.com>
Patchwork-id: 285528
O-Subject: [RHEL-8.2.0 BZ1719676 V2 41/72] IB/{hfi1, qib, rdmavt}: Put qp in error state when cq is full
Bugzilla: 1719676
RH-Acked-by: Kamal Heib <kheib@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Jonathan Toppins <jtoppins@redhat.com>

From: Kamenee Arumugam <kamenee.arumugam@intel.com>

When a completion queue is full, the associated queue pairs are not put
into the error state. According to the IBTA specification, this is a
violation.

Quote from IBTA spec:
C9-218: A Requester Class F error occurs when the CQ is inaccessible or
full and an attempt is made to complete a WQE.  The Affected QP shall be
moved to the error state and affiliated asynchronous errors generated as
described in 11.6.3.1 Affiliated Asynchronous Events on page 678. The
current WQE and any subsequent WQEs are left in an unknown state.

C11-37: The CI shall generate a CQ Error when a CQ overrun is
detected. This condition will result in an Affiliated Asynchronous Error
for any associated Work Queues when they attempt to use that
CQ. Completions can no longer be added to the CQ. It is not guaranteed
that completions present in the CQ at the time the error occurred can be
retrieved. Possible causes include a CQ overrun or a CQ protection error.

Put the qp in error state when cq is full. Implement a state called full
to continue to put other associated QPs in error state.

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Reviewed-by: Michael J. Ruhl <michael.j.ruhl@intel.com>
Signed-off-by: Kamenee Arumugam <kamenee.arumugam@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>
(cherry picked from commit 5136bfea7e79b333af77594fac5bc70282a95313)
Signed-off-by: Honggang Li <honli@redhat.com>
Signed-off-by: Phillip Lougher <plougher@redhat.com>