[infiniband] IB/hfi1: Do not update hcrc for a KDETH packet during fault injection
Message-id: <20191112031053.14620-57-honli@redhat.com>
Patchwork-id: 285543
O-Subject: [RHEL-8.2.0 BZ1719676 V2 56/72] IB/hfi1: Do not update hcrc for a KDETH packet during fault injection
Bugzilla: 1719676
RH-Acked-by: Kamal Heib <kheib@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Jonathan Toppins <jtoppins@redhat.com>

From: Kaike Wan <kaike.wan@intel.com>

When a KDETH packet is subject to fault injection during transmission,
HCRC is supposed to be omitted from the packet so that the hardware on the
receiver side would drop the packet. When creating pbc, the PbcInsertHcrc
field is set to be PBC_IHCRC_NONE if the KDETH packet is subject to fault
injection, but overwritten with PBC_IHCRC_LKDETH when update_hcrc() is
called later.

This problem is fixed by not calling update_hcrc() when the packet is
subject to fault injection.

Fixes: 6b6cf9357f78 ("IB/hfi1: Set PbcInsertHcrc for TID RDMA packets")
Cc: <stable@vger.kernel.org>
Link: https://lore.kernel.org/r/20190715164546.74174.99296.stgit@awfm-01.aw.intel.com
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Kaike Wan <kaike.wan@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>
(cherry picked from commit b2590bdd0b1dfb91737e6cb07ebb47bd74957f7e)
Signed-off-by: Honggang Li <honli@redhat.com>
Signed-off-by: Phillip Lougher <plougher@redhat.com>