[fs] autofs: add ignore mount option
Message-id: <157354618072.13649.2848404969672561578.stgit@centos8>
Patchwork-id: 285564
O-Subject: [RHEL-8.2.0 fs PATCH] autofs: add ignore mount option
Bugzilla: 1743441
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Miklos Szeredi <mszeredi@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1743441
commit 60d6d04ca3abb34d5e89f030dbea440d9715a168
Brew build: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=24604608
Testing: Used updated autofs to mount autofs mounts with the "ignore"
	option to check the kernel would accept it and display it on
	listed mounts.

Add an autofs file system mount option that can be used to provide a
generic indicator to applications that the mount entry should be ignored
when displaying mount information.

In other OSes that provide autofs and that provide a mount list to user
space based on the kernel mount list a no-op mount option ("ignore" is
the one use on the most common OS) is allowed so that autofs file system
users can optionally use it.

The idea is that it be used by user space programs to exclude autofs
mounts from consideration when reading the mounts list.

Prior to the change to link /etc/mtab to /proc/self/mounts all I needed
to do to achieve this was to use mount(2) and not update the mtab but
now that no longer works.

I know the symlinking happened a long time ago and I considered doing
this then but, at the time I couldn't remember the commonly used option
name and thought persuading the various utility maintainers would be too
hard.

But now I have a RHEL request to do this for compatibility for a widely
used product so I want to go ahead with it and try and enlist the help
of some utility package maintainers.

Clearly, without the option nothing can be done so it's at least a
start.

Link: http://lkml.kernel.org/r/154725123970.11260.6113771566924907275.stgit@pluto-themaw-net
Signed-off-by: Ian Kent <raven@themaw.net>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Ian Kent <ikent@redhat.com>
Signed-off-by: Phillip Lougher <plougher@redhat.com>