[hwtracing] intel_th: Fix resource handling for ACPI glue layer
Message-id: <20191118090749.8127-4-jolsa@redhat.com>
Patchwork-id: 287558
O-Subject: [PATCH RHEL8.2 BZ 1771015 03/46] intel_th: Fix resource handling for ACPI glue layer
Bugzilla: 1771015
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1771015
https://bugzilla.redhat.com/show_bug.cgi?id=1771015

upstream
========
commit ebe4582281d6e90972f057318a6edea14810ea48
Author: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Date: Tue Sep 18 16:10:48 2018 +0300

description
===========
The core of the driver expects the resource array from the glue layer
to be indexed by even numbers, as is the case for 64-bit PCI resources.
This doesn't hold true for others, ACPI in this instance, which leads
to an out-of-bounds access and an ioremap() on whatever address that
access fetches.

This patch fixes the problem by reading resource array differently based
on whether the 64-bit flag is set, which would indicate PCI glue layer.

Fixes: ebc57e399b8e ("intel_th: Add ACPI glue layer")
Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>