[hwtracing] intel_th: Rework resource passing between glue layers and core
Message-id: <20191118090749.8127-13-jolsa@redhat.com>
Patchwork-id: 287567
O-Subject: [PATCH RHEL8.2 BZ 1771015 12/46] intel_th: Rework resource passing between glue layers and core
Bugzilla: 1771015
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1771015
https://bugzilla.redhat.com/show_bug.cgi?id=1771015

upstream
========
commit db73a059de00eed721f13051c0d6ff3e7de90fe8
Author: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Date: Fri May 3 11:44:36 2019 +0300

description
===========
Currently, MMIO resource numbers in the TH driver core correspond to
PCI BAR numbers, because in the beginning there was only the PCI glue
layer. This created some confusion when the ACPI glue layer was added.

To avoid confusion and remove glue-specific code from the driver core,
split the resource indices between core and glue layers and change the
API so that the driver core receives the MMIO resources in the same
fixed order. At the same time, make the IRQ always be a parameter to
intel_th_alloc() instead of sometimes passing it as a resource.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>