[hwtracing] intel_th: Skip subdevices if their MMIO is missing
Message-id: <20191118090749.8127-14-jolsa@redhat.com>
Patchwork-id: 287566
O-Subject: [PATCH RHEL8.2 BZ 1771015 13/46] intel_th: Skip subdevices if their MMIO is missing
Bugzilla: 1771015
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1771015
https://bugzilla.redhat.com/show_bug.cgi?id=1771015

upstream
========
commit 23f667494b4d65d230cf61ad0d0b96c72f3a163c
Author: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Date: Fri May 3 11:44:37 2019 +0300

description
===========
If a subdevice requires an MMIO region that wasn't in the resources passed
down from the glue layer, don't instantiate it, but don't error out. This
means that that particular subdevice doesn't exist for this instance of
Intel TH, which is a perfectly normal situation. This applies, for example,
to the "rtit" source device.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>