[hwtracing] intel_th: Communicate IRQ via resource
Message-id: <20191118090749.8127-16-jolsa@redhat.com>
Patchwork-id: 287565
O-Subject: [PATCH RHEL8.2 BZ 1771015 15/46] intel_th: Communicate IRQ via resource
Bugzilla: 1771015
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1771015
https://bugzilla.redhat.com/show_bug.cgi?id=1771015

upstream
========
commit 62a593022c32380d040303a5e3d6b67fd9c415bc
Author: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Date: Fri May 3 11:44:39 2019 +0300

description
===========
Currently, the IRQ is passed between the glue layers and the core as a
separate argument, while the MMIO resources are passed as resources.
This also limits the number of IRQs thus used to one, while the current
versions of Intel TH use a different MSI vector for each interrupt
triggering event, of which there are 7.

Change this to pass IRQ in the resources array.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>