[hwtracing] intel_th: msu: Factor out pipeline draining
Message-id: <20191118090749.8127-22-jolsa@redhat.com>
Patchwork-id: 287577
O-Subject: [PATCH RHEL8.2 BZ 1771015 21/46] intel_th: msu: Factor out pipeline draining
Bugzilla: 1771015
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1771015
https://bugzilla.redhat.com/show_bug.cgi?id=1771015

upstream
========
commit 8d4155126e32fdceda7e1520e77a7beac2217f4a
Author: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Date: Fri May 3 11:44:46 2019 +0300

description
===========
The code that waits for the pipeline empty condition of the MSU is
currently called in the path that disables the trace. We will also
need this in the window switch trigger sequence. Therefore, factor
out this code and make it accessible to the GTH device.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>