[hwtracing] intel_th: msu: Add current window tracking
Message-id: <20191118090749.8127-27-jolsa@redhat.com>
Patchwork-id: 287580
O-Subject: [PATCH RHEL8.2 BZ 1771015 26/46] intel_th: msu: Add current window tracking
Bugzilla: 1771015
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1771015
https://bugzilla.redhat.com/show_bug.cgi?id=1771015

upstream
========
commit aad14ad3cf3a63bd258b65e18d49c3eb8472d344
Author: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Date: Fri May 3 11:44:51 2019 +0300

description
===========
Now that we have a way to switch between MSC buffer windows, add code to
track the current window. The hardware register NWSA that contains the
address of the next window is unfortunately not always usable, and since
the driver has full control of the window switching, there is no reason
not to keep this on the software side.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>