[hwtracing] intel_th: msu: Start read iterator from a non-empty window
Message-id: <20191118090749.8127-34-jolsa@redhat.com>
Patchwork-id: 287588
O-Subject: [PATCH RHEL8.2 BZ 1771015 33/46] intel_th: msu: Start read iterator from a non-empty window
Bugzilla: 1771015
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1771015
https://bugzilla.redhat.com/show_bug.cgi?id=1771015

upstream
========
commit f505e91ef511a20b2bc3d0eaca41faecdb715027
Author: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Date: Thu Jun 27 15:51:47 2019 +0300

description
===========
In multi-window mode, the read iterator is supposed to start from the
window with the oldest data, which is, chronologically, the next window
after the one with the newest data. This, however, fails to take into
account the potentially empty windows, so in short trace sessions it's
possible to have a lot of zeroes read from the character device first.

Fix this by skipping over the empty windows in initialization of the
read iterator.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>